2019-12-10 13:40 UTC+0300 Alexander Kresin <alex@belacy.ru>   * makefile.vc   * source/winapi/draw.c     ! Small fixes for MSVC   * source/winapi/htimer.prg   * source/winapi/control.c     ! Small fix in timer
git-svn-id: https://svn.code.sf.net/p/hwgui/code/trunk/hwgui@2798 66345640-f411-406e-acba-706ed0d45c9c

+++ source/winapi/draw.c
@@ -1477,7 +1477,7 @@ HB_FUNC( HWG_DRAWGRADIENT )
          isR = 1;
          x_center = (x2 - x1) / 2 + x1;
          y_center = (y2 - y1) / 2 + y1;
-         gr_radius = sqrt( pow(x2-x1,2) + pow(y2-y1,2) ) / 2;
+         gr_radius = sqrt( pow((long double)(x2-x1),2) + pow((long double)(y2-y1),2) ) / 2;
       }
 
       // calculate stops and colors for our gradient
@@ -1785,8 +1785,8 @@ HB_FUNC( HWG_DRAWGRADIENT )
                min_delta = 1000000;
                for( k = 0; k < 4; k++ )
                {
-                  delta = abs( pow( candidates[k].x, 2 ) + pow( candidates[k].y, 2 ) -
-                     pow( radius[i], 2 ) );
+                  delta = abs( pow( (long double)(candidates[k].x), 2 ) + pow( (long double)(candidates[k].y), 2 ) -
+                     pow( (long double)(radius[i]), 2 ) );
                   if ( delta < min_delta )
                   {
                      nearest_coord = k;
